requirements.txt
azure-common==1.1.26
azure-core==1.23.0
azure-identity==1.5.0
azure-multiapi-storage==0.6.0
azure-storage-blob==12.6.0
azure-storage-file-datalake==12.0.1
azure-graphrbac==0.61.1
azure-cli-core==2.30.0
azure-mgmt-resource==15.0.0
azure-mgmt-storage==16.0.0
azure-mgmt-datafactory==2.3.0
fastapi==0.75.0
uvicorn==0.17.6
purviewcli
python-dotenv
bandit
black
flake8
pandas
pydocstyle
pylint
pyspark
pytest-cov
pytest-mock
pytest-bdd
mkdocs
databricks-cli
databricks-sql-connector